+++ react/components/Collapsible/index.js
@@ -36,9 +36,13 @@ class Collapsible extends Component {
   }
 
   handleTransitionEnd = () => {
+    const { isOpen, isOverflowHidden } = this.props
     this.setState({
-      height: this.props.isOpen ? 'auto' : 0,
+      height: isOpen ? 'auto' : 0,
     })
+    if (isOpen && !isOverflowHidden) {
+      this.childrenRef.current.style.removeProperty('overflow')
+    }
   }
 
   openCard = () => {
@@ -58,6 +62,7 @@ class Collapsible extends Component {
 
   closeCard = () => {
     const childrenHeight = this.childrenRef.current.offsetHeight
+    this.childrenRef.current.style.setProperty('overflow', 'hidden')
     this.setState(
       {
         height: childrenHeight,
@@ -93,12 +98,13 @@ class Collapsible extends Component {
       onClick: callback,
       isOpen,
       arrowAlign,
+      isOverflowHidden,
     } = this.props
     let { caretColor } = this.props
     const { height } = this.state
     const childrenContainerStyle = {
       height,
-      overflow: 'hidden',
+      overflow: !isOverflowHidden && isOpen ? 'none' : 'hidden',
       transition: 'height 250ms ease-in-out',
     }
     if (muted) {
@@ -152,6 +158,7 @@ Collapsible.defaultProps = {
   isOpen: false,
   muted: false,
   arrowAlign: 'center',
+  isOverflowHidden: true,
 }
 
 Collapsible.propTypes = {
@@ -180,6 +187,8 @@ Collapsible.propTypes = {
     'baseline',
     'stretch',
   ]),
+  /** Controls whether the collapsible should hide overflowing components. (e.g. Turn the overflow off to avoid popup menus in the childen component to be cropped.)*/
+  isOverflowHidden: PropTypes.bool,
 }
 
 export default Collapsible