+++ react/playground/stories.tsx
@@ -1,0 +1,9 @@
+import Playground from './Playground'
+import ExamplePage from './ExamplePage'
+import All from './All'
+
+export default {
+  title: 'Playground/Playground',
+}
+
+export { Playground, ExamplePage, All }